requirements/requirement.txt
pandas
numpy
matplotlib.pyplot
request
luigi
sqlalchemy
python-dotenv